pyproject.toml
[project]
name = "streamer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faker>=33.3.0",
    "kafka-python>=2.0.2",
    "numpy>=2.2.1",
    "pydantic-settings>=2.7.1",
    "pydantic>=2.10.5",
    "pyspark>=3.5.4",
    "redis>=5.2.1",
]

[dependency-groups]
dev = [
    "black>=24.10.0",
    "ruff>=0.8.6",
]
